Fixed notification call
